+++ Federation/src/Domain/Options/PersistentGrantOptions.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-namespace Wangkanai.Federation.Stores;
+namespace Wangkanai.Federation.Options;
 
 /// <summary>Options for how persisted grants are persisted.</summary>
 public sealed class PersistentGrantOptions
