+++ Federation/src/Federation/DependencyInjection/Options/FederationOptions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
 using Wangkanai.Federation.Hosting;
-using Wangkanai.Federation.Stores;
+using Wangkanai.Federation.Options;
 
 namespace Wangkanai.Federation;
